requirements.txt
streamlit>=1.22.0
yfinance>=0.2.12
pandas>=1.5.3
numpy>=1.23.5
plotly>=5.14.1
torch>=2.0.0
chronos-forecasting>=1.5.0
pandas-datareader>=0.10.0
scikit-learn>=1.2.2
requests>=2.28.0
python-dotenv>=1.0.0
transformers>=4.30.0
safetensors>=0.3.1
accelerate>=0.20.0
optimum>=1.10.0
